confirm delete

+++ spring/angular/taskProjectAprel11/src/app/task-list/task-list.component.ts
@@ -30,10 +30,13 @@ export class TaskListComponent implements OnInit {
   }
 
   deleteTaskById(id:number){
-    this.http.delete(API_URL+'/tasks/'+id).subscribe(
-      resp=>{
-        this.loadTasks();
-      }
-    );
+    if (confirm("silmeye eminsen")) {
+      this.http.delete(API_URL+'/tasks/'+id).subscribe(
+        resp=>{
+          this.loadTasks();
+        }
+      );
+    }
+  
   }
 }
